pyproject.toml
[tool.poetry]
name = "collect"
version = "0.1.0"
description = ""
authors = ["Renntor <denis30510@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.6"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
celery = "^5.5.3"
redis = "^6.4.0"
drf-spectacular = "^0.28.0"
djoser = "^2.3.3"
flake8 = "^7.3.0"
pillow = "^11.3.0"
django-redis = "^6.0.0"
isort = "^6.1.0"
pytest = "^8.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
